Definition of _notifiables_ in ObservableObject subclasses made optional.
If _notifiables_ is not defined, any attribute set by the object not
starting with underscore fires an event

Rationale: Usually events for all regular attributes is expected.
Class attributes and/or attributes not expected to be modified don't care.
Defining _notifiables_ is burdensome for more than a few attributes per class